refactor : MenuLayout core:designsystem 모듈로 이동
* template 패키지 : 단순 반복 레이아웃을 분리해둔 것
* component 패키지 : 일반적인 컴포넌트들에 대한 구현
    -  ex) 버튼, 바텀시트, 다이얼로그, 이미지, 동영상 플레이어 등등
